Simplify pip packaging and remove cruft files
+ file setup.py had been placing the following files and directories
  in improper places as follows:

  + installed directory doc to ~/.local/lib/python3.10/site-packages/

  + installed executable nerd-dictation to
    ~/.local/lib/python3.10/site-packages/

  + installed directory nerd-dictation to ~/.local/lib/

  + installed executable and readme.rst to ~/.local/nerd-dictation/

+ file setup.py had NOT been performing the following:

  + installing directory doc to ~/.local/share/doc/nerd-dictation

    + installing documentation beyond readme.rst

    + installing example nerd-dictation.py files

  + installing executable nerd-dictation to ~/.local/bin

+ file setup.py suffered from needless complexity due to being placed
  in a sub folder instead of the based directory as is standard and
  expected practice

+ file pyproject.toml did not contain the minimum content as
  prescribed by package setuptools' documentation

  + this file exists only for the purpose of setuptools, so it was
    incongruous that it existed as it should, in the package root directory,
    even though the setup.py script to actually run setuptools was
    hidden in a sub-sub-folder.

+ a config file nerd-dictation.py is described as default, but had not been
  installed by default, so it wasn't really the default.
