--- package/python/readme.rst
@@ -1,11 +1,0 @@
-##################
-Packaging with PIP
-##################
-
-To install nerd-dictaiton as a package from pip, run:
-
-.. code-block:: sh
-
-   pip install package/python
-
-On a typical Linux system this will install nerd-dictation to ``~/.local/bin``.
