+++ package/readme.rst
@@ -5,5 +5,10 @@ Packaging
 While nerd-dictation is intended to run without the need for a build/packaging step,
 some users may find it convenient.
 
-- `Python's pip (setup-tools) <python/readme.rst>`_.
+- Python's pip (setup-tools)
+
+From the parent of the project directory, run:
+
+    pip3 install ./nerd-dictation
+
 - `GNU Guix <guix/readme.rst>`_.
